+++ scripts/config/local/configureOpenVSCodeServer.sh
@@ -1,0 +1,3 @@
+#!/bin/sh
+
+~/.local/openvscode-server/latest/server.sh &>> ~/.local/openvscode-server/logs/openvscode-server.log &